# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DX-Ray API - Developer Experience Diagnostic Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type RepositoryPlatform =
  (typeof RepositoryPlatform)[keyof typeof RepositoryPlatform];

export const RepositoryPlatform = {
  github: "github",
  gitlab: "gitlab",
  demo: "demo",
} as const;

export interface Repository {
  id: number;
  name: string;
  url: string;
  description?: string;
  language?: string;
  stars?: number;
  lastScanAt?: string;
  dxScore: number;
  platform?: RepositoryPlatform;
  repoOwner?: string;
  repoName?: string;
  isRealData?: boolean;
  createdAt: string;
}

export type AddRepositoryRequestPlatform =
  (typeof AddRepositoryRequestPlatform)[keyof typeof AddRepositoryRequestPlatform];

export const AddRepositoryRequestPlatform = {
  github: "github",
  gitlab: "gitlab",
  demo: "demo",
} as const;

export interface AddRepositoryRequest {
  /** Full repository URL (e.g. https://github.com/owner/repo) */
  url: string;
  name: string;
  /** GitHub or GitLab personal access token */
  accessToken?: string;
  platform?: AddRepositoryRequestPlatform;
}

export type ScanResultStatus =
  (typeof ScanResultStatus)[keyof typeof ScanResultStatus];

export const ScanResultStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface ScanResult {
  id: number;
  repositoryId: number;
  status: ScanResultStatus;
  tracks: string[];
  startedAt: string;
  completedAt?: string;
  summary?: string;
  dxScoreBefore?: number;
  dxScoreAfter?: number;
  isRealData?: boolean;
}

export type TriggerScanRequestTracksItem =
  (typeof TriggerScanRequestTracksItem)[keyof typeof TriggerScanRequestTracksItem];

export const TriggerScanRequestTracksItem = {
  "ci-cd": "ci-cd",
  "test-health": "test-health",
  "code-quality": "code-quality",
  "pr-review": "pr-review",
  docs: "docs",
  dependencies: "dependencies",
  onboarding: "onboarding",
  environment: "environment",
} as const;

export interface TriggerScanRequest {
  repositoryId: number;
  tracks: TriggerScanRequestTracksItem[];
}

export type TrackSummaryStatus =
  (typeof TrackSummaryStatus)[keyof typeof TrackSummaryStatus];

export const TrackSummaryStatus = {
  critical: "critical",
  warning: "warning",
  good: "good",
} as const;

export interface TrackSummary {
  id: string;
  name: string;
  score: number;
  status: TrackSummaryStatus;
  topIssue?: string;
}

export interface MetricsOverview {
  dxScore: number;
  dxScoreDelta: number;
  criticalIssues: number;
  warnings: number;
  improvements: number;
  isRealData?: boolean;
  lastFetchedAt?: string;
  tracks: TrackSummary[];
}

export interface TimeSeriesPoint {
  date: string;
  value: number;
  label?: string;
}

export interface PipelineStep {
  name: string;
  avgDurationMinutes: number;
  failureRate: number;
  isBottleneck: boolean;
}

export interface DayMetric {
  day: string;
  avgMinutes: number;
  builds: number;
}

export interface CiCdMetrics {
  avgBuildTimeMinutes: number;
  avgBuildTimeDelta: number;
  successRate: number;
  flakiness: number;
  isRealData?: boolean;
  buildTimeTrend: TimeSeriesPoint[];
  slowestSteps: PipelineStep[];
  buildsByDayOfWeek: DayMetric[];
}

export interface FailurePattern {
  pattern: string;
  count: number;
  affectedTests: number;
  isFlaky: boolean;
}

export type ModuleCoverageStatus =
  (typeof ModuleCoverageStatus)[keyof typeof ModuleCoverageStatus];

export const ModuleCoverageStatus = {
  good: "good",
  warning: "warning",
  critical: "critical",
} as const;

export interface ModuleCoverage {
  module: string;
  coverage: number;
  status: ModuleCoverageStatus;
}

export interface TestHealthMetrics {
  totalTests: number;
  flakyTests: number;
  flakyRate: number;
  coveragePercent: number;
  coverageDelta: number;
  isRealData?: boolean;
  failurePatterns: FailurePattern[];
  coverageByModule: ModuleCoverage[];
}

export interface BugPattern {
  pattern: string;
  occurrences: number;
  category: string;
  suggestedFix: string;
}

export type CodeIssueSeverity =
  (typeof CodeIssueSeverity)[keyof typeof CodeIssueSeverity];

export const CodeIssueSeverity = {
  error: "error",
  warning: "warning",
  info: "info",
} as const;

export interface CodeIssue {
  file: string;
  line?: number;
  severity: CodeIssueSeverity;
  rule: string;
  message: string;
}

export interface CodeQualityMetrics {
  overallScore: number;
  typeSafetyScore: number;
  complexityAvg: number;
  duplicateCodePercent: number;
  isRealData?: boolean;
  bugPatterns: BugPattern[];
  qualityTrend: TimeSeriesPoint[];
  topIssues: CodeIssue[];
}

export interface SizeDistributionItem {
  size: string;
  count: number;
  avgReviewTimeHours: number;
}

export interface ReviewerLoad {
  reviewer: string;
  prCount: number;
  avgResponseHours: number;
}

export interface PrReviewMetrics {
  avgTimeToFirstReviewHours: number;
  avgTimeToMergeHours: number;
  avgReviewCycles: number;
  isRealData?: boolean;
  prSizeDistribution: SizeDistributionItem[];
  reviewerLoad: ReviewerLoad[];
  reviewTimeTrend: TimeSeriesPoint[];
}

export interface DocFreshnessItem {
  category: string;
  count: number;
}

export interface StaleDoc {
  path: string;
  lastUpdated: string;
  monthsOld: number;
  category: string;
}

export interface DocsMetrics {
  totalDocs: number;
  staleDocsPercent: number;
  avgAgeMonths: number;
  missingDocs: number;
  isRealData?: boolean;
  docsByFreshness: DocFreshnessItem[];
  mostStale: StaleDoc[];
}

export type InsightSeverity =
  (typeof InsightSeverity)[keyof typeof InsightSeverity];

export const InsightSeverity = {
  critical: "critical",
  warning: "warning",
  info: "info",
} as const;

export interface Insight {
  id: number;
  track: string;
  severity: InsightSeverity;
  title: string;
  description: string;
  impact?: string;
  recommendation: string;
  estimatedHoursSaved?: number;
}

export interface BeforeAfterMetric {
  label: string;
  before: number;
  after: number;
  unit: string;
  improvement: number;
  isPositiveWhenIncreasing: boolean;
}

export interface BeforeAfterComparison {
  period: string;
  metrics: BeforeAfterMetric[];
}

export type ListScansParams = {
  repositoryId?: number;
};

export type GetMetricsOverviewParams = {
  repositoryId?: number;
};

export type GetCiCdMetricsParams = {
  repositoryId?: number;
  days?: number;
};

export type GetTestHealthMetricsParams = {
  repositoryId?: number;
};

export type GetCodeQualityMetricsParams = {
  repositoryId?: number;
};

export type GetPrReviewMetricsParams = {
  repositoryId?: number;
};

export type GetDocsMetricsParams = {
  repositoryId?: number;
};

export type ListInsightsParams = {
  repositoryId?: number;
  severity?: ListInsightsSeverity;
};

export type ListInsightsSeverity =
  (typeof ListInsightsSeverity)[keyof typeof ListInsightsSeverity];

export const ListInsightsSeverity = {
  critical: "critical",
  warning: "warning",
  info: "info",
} as const;

export type GetBeforeAfterComparisonParams = {
  repositoryId?: number;
};