# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DX-Ray API - Developer Experience Diagnostic Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /repositories:
    get:
      operationId: listRepositories
      tags: [repositories]
      summary: List repositories
      responses:
        "200":
          description: List of repositories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repository"
    post:
      operationId: addRepository
      tags: [repositories]
      summary: Add a repository to analyze
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRepositoryRequest"
      responses:
        "201":
          description: Repository added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
        "400":
          description: Invalid repository or token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /repositories/{id}:
    get:
      operationId: getRepository
      tags: [repositories]
      summary: Get repository by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Repository details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
    delete:
      operationId: deleteRepository
      tags: [repositories]
      summary: Remove a repository
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Repository:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        description:
          type: string
        language:
          type: string
        stars:
          type: integer
        lastScanAt:
          type: string
          format: date-time
        dxScore:
          type: number
        platform:
          type: string
          enum: [github, gitlab, demo]
        repoOwner:
          type: string
        repoName:
          type: string
        isRealData:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - url
        - dxScore
        - createdAt
    AddRepositoryRequest:
      type: object
      properties:
        url:
          type: string
          description: Full repository URL (e.g. https://github.com/owner/repo)
        name:
          type: string
        accessToken:
          type: string
          description: GitHub or GitLab personal access token
        platform:
          type: string
          enum: [github, gitlab, demo]
      required:
        - url
        - name
...