# lib/api-zod/src/generated/types/repository.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DX-Ray API - Developer Experience Diagnostic Platform
 * OpenAPI spec version: 0.1.0
 */
import type { RepositoryPlatform } from "./repositoryPlatform";

export interface Repository {
  id: number;
  name: string;
  url: string;
  description?: string;
  language?: string;
  stars?: number;
  lastScanAt?: Date;
  dxScore: number;
  platform?: RepositoryPlatform;
  repoOwner?: string;
  repoName?: string;
  isRealData?: boolean;
  createdAt: Date;
}